add the rating system in backend as well as in the frontend
